capitalizing filenames
